FrackMan finished project
Project completed in UCLA course number CS32. The object of this class
is to learn how to structure many classes on multiple levels using
inheritance.

The object of the game as the FrackMan is to collect all of the oil
without getting hurt by the protestors. Protestors have 5 health, and
Hardcore Protestors have 20 health. You can take away their health by
using a squirt gun. Radars reveal objects a certain radius away from
the player, and Gold can be used to bribe a Protestor (note: Hardcore
protestors will not go away if bribed, they will just be stunned)

I earned an A for my work on this project as well as an A in the course